+++ sv-parser-pp/testcases/test10.sv
@@ -1,0 +1,2 @@
+module and_op (a, b, c);
+`include "test2.svh" endmodule