+++ sv-parser-pp/testcases/test9.sv
@@ -1,0 +1,3 @@
+module and_op (a, b, c);
+`include "test2.svh"   `include "test2.svh"
+endmodule